% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Bias estimation function}
\usage{
bias(actual, predicted)
}
\arguments{
\item{actual}{vector of measured values}

\item{predicted}{vector of modeled values}
}
\value{
return bias
}
\description{
Function calculating bias
}
