% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_rmse.R
\name{bias_rmse}
\alias{bias_rmse}
\title{Test function}
\usage{
bias_rmse(x_test, y_test, y_pred)
}
\arguments{
\item{x}{data source pro processed via \link{expand_data_set}}
}
\value{
return bias and rmse
}
\description{
Function to test if all fields are working. It runs every field once to check if the pre processing is working
}
