% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Calibration}
\usage{
calibration(data_source, stopval = 0.01, maxeval = 100, ub = 1.5, lb = 0.3)
}
\arguments{
\item{data_source}{data source pro processed via \link{expand_data_set}}

\item{stopval}{Finishing criteria by error size}

\item{maxeval}{Finihsing criteria by the number of algorithmic iterations}

\item{ub}{upper boundary}

\item{lb}{lower boundary}
}
\value{
return all_metrics:
}
\description{
Function that performs the calibration of RothC for the data set input
}
\author{
Marcos Alves
}
