% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tc_per_ha.R
\name{convert_to_tc_per_ha}
\alias{convert_to_tc_per_ha}
\title{Convert Values to tC/ha}
\usage{
convert_to_tc_per_ha(unit_measure, value, sampling_depth, bulk_density = NULL)
}
\arguments{
\item{unit_measure}{The unit of measure for the value.}

\item{value}{The numeric value to be converted.}

\item{sampling_depth}{The depth in which the soil was sampled.}

\item{bulk_density}{Optional. The bulk density in kg/m^2. Required for certain unit measures.}
}
\value{
The converted value in tC/ha (metric tons of carbon per hectare).
}
\description{
This function takes a unit measure and a corresponding value and converts
the value to tC/ha (metric tons of carbon per hectare) based on the provided
unit measure and optional bulk density.
}
\examples{
# Example usage
unit_measure <- "kg/ha"
value <- 5000
bulk_density <- 1500 # kg/m^2

result <- convert_to_tc_per_ha(unit_measure, value, bulk_density)
print(result)

}
