% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tc_per_ha_fert.R
\name{convert_to_tc_per_ha_fert}
\alias{convert_to_tc_per_ha_fert}
\title{Convert Original Values to tC/ha for Specific Fertilizer Types}
\usage{
convert_to_tc_per_ha_fert(data)
}
\arguments{
\item{data}{A dataframe containing the original values, original unit measures, and fertilizer types.}
}
\value{
A modified dataframe with the converted values in tC/ha.
}
\description{
This function takes a dataframe \code{data} as input and converts the values
in the "values" column from the "original_unit" and "fert_id"
to tC/ha. It considers specific fertilizer types and their corresponding conversion
factors and carbon contents.
}
\examples{
data <- data.frame(
  values = c("2240, 2358", "2240, 2358, 2035"),
  original_unit = c("kg ha-1 yr-1", "kg ha-1 yr-1"),
  fert_id = c("manure", "compost")
)
converted_data <- convert_to_tc_per_ha_fert(data)
print(converted_data)

}
