% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_type.R
\name{fix_column_type}
\alias{fix_column_type}
\title{Fix Column Type}
\usage{
fix_column_type(data)
}
\arguments{
\item{data}{The input dataframe.}
}
\value{
The dataframe with columns converted to type character.
}
\description{
Converts columns of unknown type containing characters, NULL, NA, and list values
into columns of type character in a dataframe.
}
\examples{
# Example dataframe
data <- data.frame(
  id = 1:5,
  values = list(letters[1:3], NULL, NA, letters[4:5], character()),
  other_column = c("text", NULL, "NA", "more text", "")
)

# Convert all columns in the dataframe
converted_data <- fix_column_type(data)

# Output the converted dataframe
print(converted_data)

}
