% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tc_per_ha_fert.R
\name{get_carbon_content}
\alias{get_carbon_content}
\title{Get Carbon Content for Fertilizer Type}
\usage{
get_carbon_content(fert_id)
}
\arguments{
\item{fert_id}{The type of organic fertilizer.}
}
\value{
The carbon content for the specified fertilizer type.
}
\description{
This function retrieves the carbon content for a specific fertilizer type from the organic_fert_composition table.
}
