% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tc_per_ha_fert.R
\name{get_conversion_factor}
\alias{get_conversion_factor}
\title{Get Conversion Factor for Unit Measure}
\usage{
get_conversion_factor(unit_measure)
}
\arguments{
\item{unit_measure}{The original unit measure.}
}
\value{
The conversion factor for the specified unit measure.
}
\description{
This function retrieves the conversion factor for a specific unit measure from the unit_conversion table.
}
