% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joined_stats.R
\name{joined_stats}
\alias{joined_stats}
\title{Join Bias, RMSE, Confidence Interval, and PMU values}
\usage{
joined_stats(
  bias_pooled_comb,
  rmse_pooled_comb,
  studycount,
  Cis,
  PMU,
  join_cols
)
}
\arguments{
\item{bias_pooled_comb}{The dataframe containing bias values.}

\item{rmse_pooled_comb}{The dataframe containing RMSE values.}

\item{studycount}{The dataframe containing study counts.}

\item{Cis}{The dataframe containing confidence interval values.}

\item{PMU}{The dataframe containing PMU values.}

\item{join_cols}{Character vector of column names to join the dataframes on.}
}
\value{
A dataframe containing the merged data from all the left joins.
}
\description{
This function performs left joins on several dataframes to calculate Bias, RMSE, Confidence Interval, and PMU values.
}
\examples{
bias_rmse_ci_pmu <- joined_stats(bias_pooled_comb, rmse_pooled_comb, studycount, Cis, PMU, c("Practice_Category", "Climate_zones_IPCC", "CFGs"))
}
