% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optm_fun.R
\name{optm_fun}
\alias{optm_fun}
\title{Optimization function}
\usage{
optm_fun(params, train_data)
}
\arguments{
\item{params}{vector of calibration coeficients}

\item{x_train}{input data.frame}

\item{y_train}{validation input}
}
\value{
return mse
}
\description{
Function describing the opmization goal
}
