% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validt.R
\name{plot_validt}
\alias{plot_validt}
\title{Test function}
\usage{
plot_validt(Cis, plots_path)
}
\arguments{
\item{x}{data source pro processed via \link{expand_data_set}}
}
\value{
return bias
}
\description{
Function to test if all fields are working. It runs every field once to check if the pre processing is working
}
