% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_validt.R
\name{plot_validt3}
\alias{plot_validt3}
\title{Plot Validation Results}
\usage{
plot_validt(Cis, plots_path)
}
\arguments{
\item{Cis}{A list containing two data frames: Ci_percent and Ci. These data frames
are the output of the \code{conf_int} function.}

\item{plots_path}{Character string specifying the path where the generated plots will be saved.}
}
\value{
A list of ggplot objects representing the validation plots.
}
\description{
This function generates a set of validation plots for the confidence intervals.
}
\details{
The function generates multiple plots for each group in the data.
Three types of plots are generated:
\enumerate{
\item Scatter plot comparing the measured and modeled values.
\item Histogram of the residual errors.
\item Plot showing the modeled effect with confidence intervals.
}
}
\examples{
# Example usage with the output of conf_int function
Cis <- list(
  Ci_percent = data.frame(Practice_Category = c("A", "B"), CFGs = c("CFG1", "CFG2"),
                          Climate_zones_IPCC = c("Zone1", "Zone2"),
                          ci_percent = c(80, 60), fields_inside = c(4, 3),
                          total_fields = c(5, 5), addtional_fields_to_90 = c(1, 2)),
  Ci = data.frame(Practice_Category = c("A", "A", "B", "B"),
                  CFGs = c("CFG1", "CFG2", "CFG1", "CFG2"),
                  Climate_zones_IPCC = c("Zone1", "Zone1", "Zone2", "Zone2"),
                  observed_diff = c(10, 12, 15, 8),
                  predicted_diff = c(9, 11, 13, 6),
                  model_error = c(-1, -1, -2, -2),
                  sd = c(0.5, 0.7, 0.8, 0.6),
                  n = c(3, 4, 5, 6),
                  cilower = c(8, 9, 11, 6),
                  ciupper = c(12, 13, 17, 10),
                  fill = c("Outside C.I.", "Inside C.I.", "Outside C.I.", "Inside C.I."))
)
plots_path <- "path/to/save/plots"
plot_validt(Cis, plots_path)

}
\seealso{
See \code{\link{conf_int}} for calculating the confidence intervals.
}
