% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulations}
\alias{simulations}
\title{Simulations function}
\usage{
simulations(params, test_data)
}
\arguments{
\item{params}{vector of calibration coeficients}

\item{test_data}{data frame containing the data to be run in test}
}
\value{
return mse
}
\description{
Function to run simulations with the calibration factors and test data
}
