% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_units.R
\name{standardize_units}
\alias{standardize_units}
\title{Convert Units to Standardized Units}
\usage{
standardize_units(unit)
}
\arguments{
\item{unit}{A character string specifying the original unit.}
}
\value{
The conversion factor to convert the given unit to its standardized equivalent.
}
\description{
This function converts units to their standardized equivalents based on the unit_conversion dataset.
}
\keyword{conversion}
\keyword{standardized}
\keyword{unit}
