% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_source.R
\name{test_data_source}
\alias{test_data_source}
\title{Test each row in data_source using ag_rothc()}
\usage{
test_data_source(data_source)
}
\arguments{
\item{data_source}{A data frame containing the rows to be tested.}
}
\value{
A vector containing the "Field Id" values for rows that caused errors when passed to ag_rothc().
}
\description{
This function tests each row in data_source by feeding it into ag_rothc().
When an error occurs, the corresponding value from the "Field Id" column is stored in a vector and returned at the end.
}
