% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fields.R
\name{test_fields}
\alias{test_fields}
\title{Test function}
\usage{
test_fields(x)
}
\arguments{
\item{x}{data source pro processed via \link{expand_data_set}}
}
\value{
return bias
}
\description{
Function to test if all fields are working. It runs every field once to check if the pre processing is working
}
