% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Test function}
\usage{
conf_int(data_out)
}
\arguments{
\item{x}{data source pro processed via \link{expand_data_set}}
}
\value{
return bias
}
\description{
Function to test if all fields are working. It runs every field once to check if the pre processing is working
}
